requirements.txt
streamlit
python-dotenv
pypdf2
pandas
groq
langchain
chromadb
gspread
oauth2client